fix: resolve economy truncation, UI state desyncs, and implement Admin UI
- Database: Added V18 migration to upgrade pending resource columns to DOUBLE PRECISION, preventing fractional production loss.
- Backend (Economy): Refactored EconomyService to utilize double precision for accrual and enforce Keep 'maxStorage' limits during resource claims.
- Backend (DTO): Applied @JsonProperty mapping to the isAdmin flag to ensure consistent JSON serialization for the frontend.
- Frontend (State): Refactored KingdomView to use type-based selection (selectedGroupType) rather than stale objects, ensuring the UI refreshes immediately upon construction completion.
- Frontend (Ticker): Updated the Layout ticker and accumulator to process decimals, allowing the Keep storage bars to update smoothly in real-time even with low production rates.
- Frontend (UX): Fixed a local storage typo in LoginView preventing correct daily loot highlighting and removed hardcoded Tavern unlock logic in favor of backend config.
- Admin: Implemented the AdminView and sidebar navigation logic, gated by the authenticated isAdmin flag."

+++ src/main/java/com/keves/dreamreach/controller/PlayerController.java
@@ -63,11 +63,13 @@ public class PlayerController {
 
         PlayerProfile profile = account.getProfile();
 
+        // Run the "lazy checks" to process offline time for the economy and the Tavern
         economyService.updateProductionState(profile);
         tavernService.processArrivals(profile);
 
         PlayerPopulation pop = profile.getPopulation();
 
+        // Calculate actual keep level
         int keepLevel = profile.getBuildings().stream()
                 .filter(b -> b.getBuildingType().equalsIgnoreCase("keep"))
                 .mapToInt(BuildingInstance::getLevel)
@@ -83,6 +85,7 @@ public class PlayerController {
         int woodRate = (pop != null ? pop.getWoodcutters() * economyConfig.getWoodPerWoodcutter() : 0) + economyConfig.getBasePassiveWood();
         int stoneRate = (pop != null ? pop.getStoneworkers() * economyConfig.getStonePerStoneworker() : 0) + economyConfig.getBasePassiveStone();
 
+        // Map real database building instances to Response DTOs
         List<BuildingInstanceResponse> buildingResponses = profile.getBuildings().stream()
                 .map(b -> BuildingInstanceResponse.builder()
                         .id(b.getId())
@@ -185,10 +188,11 @@ public class PlayerController {
                 .woodRate(woodRate)
                 .stoneRate(stoneRate)
 
-                .pendingFood(profile.getResources() != null ? (int) profile.getResources().getPendingFood() : 0)
-                .pendingWood(profile.getResources() != null ? (int) profile.getResources().getPendingWood() : 0)
-                .pendingStone(profile.getResources() != null ? (int) profile.getResources().getPendingStone() : 0)
-                .pendingGold(profile.getResources() != null ? (int) profile.getResources().getPendingGold() : 0)
+                // Pass the doubles directly to allow the frontend ticker to process fractions
+                .pendingFood(profile.getResources() != null ? profile.getResources().getPendingFood() : 0)
+                .pendingWood(profile.getResources() != null ? profile.getResources().getPendingWood() : 0)
+                .pendingStone(profile.getResources() != null ? profile.getResources().getPendingStone() : 0)
+                .pendingGold(profile.getResources() != null ? profile.getResources().getPendingGold() : 0)
 
                 .happiness(profile.getHappiness())
                 .maxHappiness(economyConfig.getMaxHappiness())
